refactor to create values package
