+++ docs/nitpick-exceptions
@@ -28,6 +28,7 @@ py:class loguru.Logger
 # Type aliases don't play nicely with sphinx when you import them
 # elsewhere.  Works fine for static type checker though.  I think this
 # is because sphinx does runtime checks.
+py:class NumberLike
 py:class NumericArray
 py:class ClockTime
 py:class Time
