--- docs/source/api_reference/framework/values.rst
@@ -1,1 +1,0 @@
-.. automodule:: vivarium.framework.values
