+++ docs/source/api_reference/framework/values/pipeline.rst
@@ -1,0 +1,1 @@
+.. automodule:: vivarium.framework.values.pipeline
