+++ docs/source/tutorials/boids.rst
@@ -256,7 +256,7 @@ You can find an overview of the values system :ref:`here <values_concept>`.
 
 The Builder class exposes an additional property for working with value pipelines:
 :meth:`vivarium.framework.engine.Builder.value`.
-We call the :meth:`vivarium.framework.values.ValuesInterface.register_value_producer`
+We call the :meth:`vivarium.framework.values.manager.ValuesInterface.register_value_producer`
 method to register a new pipeline.
 
 .. literalinclude:: ../../../src/vivarium/examples/boids/movement.py
@@ -474,7 +474,7 @@ To access the value pipeline we created in the Neighbors component, we use
 pipeline, we simply call that pipeline as a function inside ``on_time_step`` to retrieve
 its values for a specified index.
 The major new Vivarium feature seen here is that of the **value modifier**,
-which we register with :meth:`vivarium.framework.values.ValuesInterface.register_value_modifier`.
+which we register with :meth:`vivarium.framework.values.manager.ValuesInterface.register_value_modifier`.
 As the name suggests, this allows us to modify the values in a pipeline,
 in this case adding the effect of a force to the values in the ``acceleration`` pipeline.
 We register that the ``apply_force`` method will modify the acceleration values like so:
