+++ src/vivarium/framework/values/exceptions.py
@@ -1,0 +1,7 @@
+from vivarium.exceptions import VivariumError
+
+
+class DynamicValueError(VivariumError):
+    """Indicates an improperly configured value was invoked."""
+
+    pass
