+++ src/vivarium/framework/values/manager.py
@@ -1,349 +1,28 @@
-"""
-=========================
-The Value Pipeline System
-=========================
-
-The value pipeline system is a vital part of the :mod:`vivarium`
-infrastructure. It allows for values that determine the behavior of individual
-:term:`simulants <Simulant>` to be constructed across multiple
-:ref:`components <components_concept>`.
-
-For more information about when and how you should use pipelines in your
-simulations, see the value system :ref:`concept note <values_concept>`.
-
-"""
 from __future__ import annotations
 
 import warnings
 from collections.abc import Callable, Iterable, Sequence
-from datetime import timedelta
-from typing import TYPE_CHECKING, Any, Protocol, TypeVar
+from typing import TYPE_CHECKING, Any, TypeVar
 
-import pandas as pd
-
-from vivarium.exceptions import VivariumError
 from vivarium.framework.event import Event
 from vivarium.framework.randomness import RandomnessStream
 from vivarium.framework.resource import Resource
-from vivarium.framework.utilities import from_yearly
+from vivarium.framework.values.combiners import replace_combiner
+from vivarium.framework.values.pipeline import (
+    MissingValueSource,
+    Pipeline,
+    PostProcessor,
+    ValueCombiner,
+    ValueModifier,
+    ValueSource,
+)
+from vivarium.framework.values.post_processors import rescale_post_processor
 from vivarium.manager import Interface, Manager
-from vivarium.types import NumberLike
 
 if TYPE_CHECKING:
     from vivarium.framework.engine import Builder
 
 T = TypeVar("T")
-
-
-class ValueCombiner(Protocol):
-    def __call__(
-        self, value: Any, mutator: Callable[..., Any], *args: Any, **kwargs: Any
-    ) -> Any:
-        ...
-
-
-class PostProcessor(Protocol):
-    def __call__(self, value: Any, manager: ValuesManager) -> Any:
-        ...
-
-
-class DynamicValueError(VivariumError):
-    """Indicates an improperly configured value was invoked."""
-
-    pass
-
-
-def replace_combiner(
-    value: Any, mutator: Callable[..., Any], *args: Any, **kwargs: Any
-) -> Any:
-    """Replace the previous pipeline output with the output of the mutator.
-
-    This is the default combiner.
-
-    Parameters
-    ----------
-    value
-        The value from the previous step in the pipeline.
-    mutator
-        A callable that takes in all arguments that the pipeline source takes
-        in plus an additional last positional argument for the value from
-        the previous stage in the pipeline.
-    args, kwargs
-        The same args and kwargs provided during the invocation of the
-        pipeline.
-
-    Returns
-    -------
-        A modified version of the input value.
-    """
-    expanded_args = list(args) + [value]
-    return mutator(*expanded_args, **kwargs)
-
-
-def list_combiner(
-    value: list[Any], mutator: Callable[..., Any], *args: Any, **kwargs: Any
-) -> list[Any]:
-    """Aggregates source and mutator output into a list.
-
-    This combiner is meant to be used with a post-processor that does some
-    kind of reduce operation like summing all values in the list.
-
-    Parameters
-    ----------
-    value
-        A list of all values provided by the source and prior mutators in the
-        pipeline.
-    mutator
-        A callable that returns some portion of this pipeline's final value.
-    args, kwargs
-        The same args and kwargs provided during the invocation of the
-        pipeline.
-
-    Returns
-    -------
-        The input list with new mutator portion of the pipeline value
-        appended to it.
-    """
-    value.append(mutator(*args, **kwargs))
-    return value
-
-
-def rescale_post_processor(value: NumberLike, manager: ValuesManager) -> NumberLike:
-    """Rescales annual rates to time-step appropriate rates.
-
-    This should only be used with a simulation using a
-    :class:`~vivarium.framework.time.DateTimeClock` or another implementation
-    of a clock that traffics in pandas date-time objects.
-
-    Parameters
-    ----------
-    value
-        Annual rates, either as a number or something we can broadcast
-        multiplication over like a :mod:`numpy` array or :mod:`pandas`
-        data frame.
-    manager
-        The ValuesManager for this simulation.
-
-    Returns
-    -------
-        The annual rates rescaled to the size of the current time step size.
-    """
-    if isinstance(value, (pd.Series, pd.DataFrame)):
-        return value.mul(
-            manager.simulant_step_sizes(value.index)
-            .astype("timedelta64[ns]")
-            .dt.total_seconds()
-            / (60 * 60 * 24 * 365.0),
-            axis=0,
-        )
-    else:
-        time_step = manager.step_size()
-        if not isinstance(time_step, (pd.Timedelta, timedelta)):
-            raise DynamicValueError(
-                "The rescale post processor requires a time step size that is a "
-                "datetime timedelta or pandas Timedelta object."
-            )
-        return from_yearly(value, time_step)
-
-
-def union_post_processor(values: list[NumberLike], _: Any) -> NumberLike:
-    """Computes a probability on the union of the sample spaces in the values.
-
-    Given a list of values where each value is a probability of an independent
-    event, this post processor computes the probability of the union of the
-    events.
-
-    .. list-table::
-       :width: 100%
-       :widths: 1 3
-
-       * - :math:`p_x`
-         - Probability of event x
-       * - :math:`1 - p_x`
-         - Probability of not event x
-       * - :math:`\prod_x(1 - p_x)`
-         - Probability of not any events x
-       * - :math:`1 - \prod_x(1 - p_x)`
-         - Probability of any event x
-
-    Parameters
-    ----------
-    values
-        A list of independent proportions or probabilities, either
-        as numbers or as a something we can broadcast addition and
-        multiplication over.
-
-    Returns
-    -------
-        The probability over the union of the sample spaces represented
-        by the original probabilities.
-    """
-    # if there is only one value, return the value
-    if len(values) == 1:
-        return values[0]
-
-    # if there are multiple values, calculate the joint value
-    product: NumberLike = 1
-    for v in values:
-        new_value = 1 - v
-        product = product * new_value
-    joint_value = 1 - product
-    return joint_value
-
-
-class ValueSource(Resource):
-    """A resource representing the source of a value pipeline."""
-
-    def __init__(self, name: str) -> None:
-        super().__init__("value_source", name)
-
-
-class MissingValueSource(Resource):
-    """A resource representing an undefined source of a value pipeline."""
-
-    def __init__(self, name: str) -> None:
-        super().__init__("missing_value_source", name)
-
-
-class ValueModifier(Resource):
-    """A resource representing a modifier of a value pipeline."""
-
-    def __init__(self, name: str) -> None:
-        super().__init__("value_modifier", name)
-
-
-class Pipeline(Resource):
-    """A tool for building up values across several components.
-
-    Pipelines are lazily initialized so that we don't have to put constraints
-    on the order in which components are created and set up. The values manager
-    will configure a pipeline (set all of its attributes) when the pipeline
-    source is created.
-
-    As long as a pipeline is not actually called in a simulation, it does not
-    need a source or to be configured. This might occur when writing
-    generic components that create a set of pipeline modifiers for
-    values that won't be used in the particular simulation.
-    """
-
-    def __init__(self, name: str) -> None:
-        super().__init__("value", name)
-
-        self.source: Callable[..., Any] | None = None
-        """The callable source of the value represented by the pipeline."""
-        self.mutators: list[Callable[..., Any]] = []
-        """A list of callables that directly modify the pipeline source or
-        contribute portions of the value."""
-        self._combiner: ValueCombiner | None = None
-        self.post_processor: PostProcessor | None = None
-        """An optional final transformation to perform on the combined output of
-        the source and mutators."""
-        self._manager: ValuesManager | None = None
-
-    def _get_attr_error(self, attribute: str) -> str:
-        return (
-            f"The pipeline for {self.name} has no {attribute}. This likely means "
-            f"you are attempting to modify a value that hasn't been created."
-        )
-
-    def _set_attr_error(self, attribute: str, new_value: Any) -> str:
-        current_value = getattr(self, f"_{attribute}")
-        return (
-            f"A second component is attempting to set the {attribute} for pipeline {self.name} "
-            f"with {new_value}, but it already has a {attribute}: {current_value}."
-        )
-
-    def _get_property(self, property: T | None, property_name: str) -> T:
-        if property is None:
-            raise DynamicValueError(self._get_attr_error(property_name))
-        return property
-
-    @property
-    def combiner(self) -> ValueCombiner:
-        """A strategy for combining the source and mutator values into the
-        final value represented by the pipeline."""
-        return self._get_property(self._combiner, "combiner")
-
-    @property
-    def manager(self) -> ValuesManager:
-        """A reference to the simulation values manager."""
-        return self._get_property(self._manager, "manager")
-
-    def __call__(self, *args: Any, skip_post_processor: bool = False, **kwargs: Any) -> Any:
-        """Generates the value represented by this pipeline.
-
-        Arguments
-        ---------
-        skip_post_processor
-            Whether we should invoke the post-processor on the combined
-            source and mutator output or return without post-processing.
-            This is useful when the post-processor acts as some sort of final
-            unit conversion (e.g. the rescale post processor).
-        args, kwargs
-            Pipeline arguments.  These should be the arguments to the
-            callable source of the pipeline.
-
-        Returns
-        -------
-            The value represented by the pipeline.
-
-        Raises
-        ------
-        DynamicValueError
-            If the pipeline is invoked without a source set.
-        """
-        return self._call(*args, skip_post_processor=skip_post_processor, **kwargs)
-
-    def _call(self, *args: Any, skip_post_processor: bool = False, **kwargs: Any) -> Any:
-        if not self.source:
-            raise DynamicValueError(
-                f"The dynamic value pipeline for {self.name} has no source. This likely means "
-                f"you are attempting to modify a value that hasn't been created."
-            )
-        value = self.source(*args, **kwargs)
-        for mutator in self.mutators:
-            value = self.combiner(value, mutator, *args, **kwargs)
-        if self.post_processor and not skip_post_processor:
-            return self.post_processor(value, self.manager)
-        if isinstance(value, pd.Series):
-            value.name = self.name
-
-        return value
-
-    def __repr__(self) -> str:
-        return f"_Pipeline({self.name})"
-
-    @classmethod
-    def setup_pipeline(
-        cls,
-        pipeline: Pipeline,
-        source: Callable[..., Any],
-        combiner: ValueCombiner,
-        post_processor: PostProcessor | None,
-        manager: ValuesManager,
-    ) -> None:
-        """
-        Add a source, combiner, and post-processor to a pipeline.
-
-        Parameters
-        ----------
-        pipeline
-            The pipeline to configure.
-        source
-            The callable source of the value represented by the pipeline.
-        combiner
-            A strategy for combining the source and mutator values into the
-            final value represented by the pipeline.
-        post_processor
-            An optional final transformation to perform on the combined output
-            of the source and mutators.
-        manager
-            The simulation values manager.
-        """
-        pipeline.source = source
-        pipeline._combiner = combiner
-        pipeline.post_processor = post_processor
-        pipeline._manager = manager
 
 
 class ValuesManager(Manager):
